requirements.txt
# Core async libraries
aiohttp>=3.8.0
aiofiles>=23.0.0

# Caching
diskcache>=5.6.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Image processing
Pillow>=10.0.0

# Progress bars
tqdm>=4.65.0

# Retry logic
tenacity>=8.2.0

# OpenAI
openai>=1.0.0

# Environment variables
python-dotenv>=1.0.0

# Type hints (for older Python versions)
typing-extensions>=4.5.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.16.0